Begin separate the femtovg renderer into a module of its own
