+++ internal/backends/gl/renderer/femtovg/images.rs
@@ -8,7 +8,7 @@ use i_slint_core::api::euclid;
 use i_slint_core::graphics::{ImageCacheKey, IntSize, SharedImageBuffer};
 use i_slint_core::{items::ImageRendering, ImageInner};
 
-use super::glrenderer::{CanvasRc, GLItemRenderer};
+use super::itemrenderer::{CanvasRc, GLItemRenderer};
 
 pub struct Texture {
     pub id: femtovg::ImageId,
